📂 Fix ShellExecute to properly determine working directory (#25)
ShellExecute commands would previously always pass a null starting
directory to the
[ShellExecuteW](https://learn.microsoft.com/en-us/windows/win32/api/shellapi/nf-shellapi-shellexecutew)
function, resulting in applications getting launched in FlashCom's
working directory, breaking apps that load binaries or other resources
from their directory on disk.

This change introduces a new optional JSON property `"executeDirectory"`
that allows a user to specify a starting directory.

If a starting directory is not provided, the specified file path is used
to determine the working directory. If the path does not provide a
parent directory, then a null value is passed to `ShellExecuteW`.