🐛 Add hotkey reset logic to avoid stuck keys (#28)
Occasionally, the hotkey management logic would get "stuck" thinking a
key is pressed when it isn't.

This can happen if keyboard events are prevented from reaching the
FlashCom process for a variety of reasons (ex. if the user invokes the
lock screen with Win+L, the key-up event for the Win key will not reach
FlashCom).

This change adds a hotkey reset timer that resets all hotkey state after
1 second of inactivity. The timer is reset each time a new hotkey key is
pressed. This will prevent FlashCom from maintaining a bad keyboard
state, even if keyboard events are interrupted.

This change also bumps the version to 0.2.1.